core-be-main/core-be-main/requirements.txt
black==23.3.0
Django==4.2.1
djangorestframework==3.14.0
drf-yasg==1.21.5
isort==5.12.0
psycopg2-binary==2.9.6
python-decouple==3.8
django-environ==0.4.5
boto3==1.19.5
Werkzeug==2.2.2
PyJWT==2.7.0
django-cors-headers
cryptography
cexprtk==0.4.1
typing-extensions==4.4.0
pyyaml==5.4.1
Brotli==1.0.9
dropbox-sign
docxtpl==0.16.6
docx2txt==0.8
requests==2.28.2
python-docx==0.8.11
